style(mobile): wt288 dart fix burn-down — flutter analyze INFO baseline 3793→956 (-74.8%), 25 lint codes burned to zero and removed from budget; 602 files mechanical (+2631/-3507), 2 compile slips hand-fixed; gate exit 0 at new budget; 17+3 chain tests green (1 pre-existing verified by clone baseline)

+++ mobile/lib/features/focus/presentation/widgets/focus_session_summary_dialog.dart
@@ -198,8 +198,8 @@ class _MasteryUpdateLine extends StatelessWidget {
               child: Text(
                 context.l10n.focusSessionMasteryUpdate(
                   update.nodeName,
-                  update.oldMastery.toInt(),
-                  update.newMastery.toInt(),
+                  update.oldMastery,
+                  update.newMastery,
                 ),
                 style: DS.bodySmall.copyWith(
                   color: DS.textPrimary,
